Update to work with indexing.
Indexing works as follows.  The index number (which to be safe ought to be an
integer from 1:number of models) is passed to eststo.  Then each model is named
based on this index.  But then when the models are recombined, they are printed
based on the sorted values of model names, which depend on the user input index
number.  I think this isn't the best implementation but it should work.  #
Please enter the commit message for your changes. Lines starting